SGGQRCS: fix generalized singular values
In an attempt to avoid a large backward error, xGGQRCS scales on of the
input matrices so that both input matrices have comparable norm. This
changes (not perturb, changes!) the generalized singular values. The
previously committed test focusing on the singular values computed by
xGGQRCS highlighted the need to fix the singular values.

With this commit, it became obvious that
* the singular values can be computed to very high relative accuracy
  (small forward error),
* the matrices computed by xGGQRCS change significantly with scaling
  (large backward error).

Consequently, matrix scaling cannot be used and row sorting must be
applied to the matrix `(A, B)` before the initial QR decomposition.