SGGQRCS: ensure matrix scaling factors always >1
The singular values must be corrected for matrix scaling. Let w be a
power of two such that norm(wA) equals approximately norm(B). Then the
radians representation of the singular values must corrected with

x' = arctan(w * tan(x)),

where x is the computed radians value with matrix scaling and x' is the
computed radians value without matrix scaling. If w is less than one
and if x is large (close to pi/2), then roughly speaking

* tan(x) has a large derivative (i.e. it is badly conditioned),
* w * tan(x) will be smaller than tan(x),
* arctan(w * tan(x)) has derivative near one, because its argument
  is closer to zero.

This commit modifies the GSVD computation such that w >= 1 by choosing
to compute the GSVD of (A, B) or (B, A) depending on the relative norms.
Then

* tan(x) has a small derivative if x is sufficiently small,
* arctan(w * tan(x)) has derivative near zero, and
* if x is large (near pi/2), then x' will be near pi/2 as well.

This commit fixes the large relative forward errors detected by
xGGQRCS_test_singular_values<float>' when the singular values of the
matrix pair (A,B) where drawn from [0, pi/2000).