Test: re-use xGGQRCS output in a test
Re-use xGGQRCS output to generate a new xGGQRCS input. The motivation is
devising a test where xGGQRCS is _not_ computing one or more matrices of
the GSVD but where one can still evaluate the backward error by
substituting the missing output matrices. For example, let A = U1 S X.
xGGQRCS should return outputs that are close in norm to U1, S, and X,
respectively; this can be evaluated with the aid of the backward error.
If either U1 or X is not computed, then one can use the original matrices
for the backward error calculation but this does not work in the
presence of (perfectly legimitate) row and column permutations. By
re-using xGGQRCS output, this commit tries to avoid this issue.