Tests: avoid tiny non-zero GSVD angles
Explicitly set generalized eigenvalue singular value pairs to (1,0) if
the angle is π/2. Otherwise the computation of the cosine yields a
non-zero value small in modulus that causes test failures in double
precision (the matrix B cannot be accurately re-assembled from the
computed GSVD). As a side effect, the matrix B can be actually zero now.